+++ src/data/mock/index.ts
@@ -5,5 +5,6 @@ export * from "./generated-course";
 export * from "./learner";
 export * from "./learning-path";
 export * from "./learning-experience";
+export * from "./organization";
 export * from "./pre-assessment";
 export * from "./product";